global functions makeOrientedBoundingBoxFromRegion() and intersectPlaneWithRay() to prevent large files and improve tree shaking
